requirements.txt
# Core dependencies
vllm>=0.6.0
openai>=1.0.0
httpx>=0.24.0
imageio[ffmpeg]>=2.30.0
Pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.24.0

# Optional but recommended for performance
uvloop>=0.19.0
accelerate>=0.26.0

# Requirements for COCO Dataset Generation with VL Model

# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# VL Model and API
openai>=1.0.0
httpx>=0.24.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Progress and logging
tqdm>=4.66.0

# Data processing
pyyaml>=6.0

# Optional: For better visualization
matplotlib>=3.7.0
